Extract bitflags_array from `mod features`
